+++ backend/requirements.txt
@@ -1,0 +1,6 @@
+fastapi
+uvicorn
+python-multipart
+pillow
+matplotlib
+numpy
