+++ frontend/main.py
@@ -6,4 +6,4 @@ def home():
     return render_template('index.html')
 
 if __name__ == '__main__':
-    app.run(debug=True)
+    app.run(host='0.0.0.0', port=8090, debug=True)
